StandaloneMmPkg: Allocate and initialise SP stack from internal memory
This patch removes the dependency on the SPM to allocate and initialise stack
memory for the StMM SP. This is done by reserving 8K worth of memory in the StMM
image at a page aligned address in the data section. Then, instead of jumping
directly to the C entrypoint, an assembler entrypoint is invoked. This
entrypoint locates the stack memory, changes its permissions using FF-A ABIs,
sets the stack pointer and invokes the C entrypoint.

Signed-off-by: Achin Gupta <achin.gupta@arm.com>
Signed-off-by: Nishant Sharma <nishant.sharma@arm.com>
Change-Id: Ic2d87de5611a4af0cfffd6fcd3eb9a63397b890d