StandaloneMmPkg: define new data structure to stage FF-A boot information
With FF-A v1.1, the SPMC sends a reduced amount of boot information as compared
to the original SPM implementation. For example, the stack layout, MP
information etc. This information could be accommodated in the old data
structure but this is too complicated. This patch defines a new structure to
clearly separate FF-A v1.1 and other functionality.

Signed-off-by: Achin Gupta <achin.gupta@arm.com>
Signed-off-by: Nishant Sharma <nishant.sharma@arm.com>
Change-Id: If0fae97197703147402299dd50e681ac4699fc25