StandaloneMmPkg: Add support to detect FF-A v1.1 in a backwards compatible manner
For better or worse, an StMM SP can communicate with the SPM through one
of these interfaces.

1. Original TF-A specific interface exported by the SPM
2. FF-A v1.0 interface
3. FF-A v1.1 interface

2) implements only minimal FF-A support. It reuses the initialisation ABI
defined by 1) and wraps the remaining communicaton in
FFA_MSG_SEND_DIRECT_REQ/RESP ABIs. 3) uses FF-A ABIs from the spec for both
initialisation and communication.

Detecting these variations in the GetSpmVersion() function is tedious. This
patch restores the original function that discovered configuration 1). It
defines a new function to discover presence of FF-A and differentiate between
v1.0 and v1.1.

Signed-off-by: Achin Gupta <achin.gupta@arm.com>
Signed-off-by: Nishant Sharma <nishant.sharma@arm.com>
Change-Id: I7e14a434623b5c6da3970df77fc4c8b1142e9ee1