FFA: Use the offset field to calculate the sp membase
With SPMC at EL2 enabled, StMM is loaded by the SP framework in BL2.
This framework packages the SP and its manifest into a single package
image. The package is signed, the information is stored in the header at
the front, and the SP image comes afterwards. The header size can
increase in the future and cannot be hardcoded. Add the support to pass
the SP image offset to avoid using hardcoded values.